added update  controller

+++ controllers/product.controller.js
@@ -14,29 +14,49 @@ const CreateProduct = async (req, res) => {
   }
 };
 
-const getAllProductsTesting = async (req, res, next) => {
-  const result = await ProductModel.find().skip(0).limit(20);
+const getAllProducts = async (req, res, next) => {
+  const result = await ProductModel.find();
   res.status(200).json({ msg: result });
   next();
 };
 
 const getById = async (req, res, next) => {
   const id = req.params.id;
-  console.log()
   const product = await ProductModel.findOne({ _id: id });
-  return res.success("Delted Successfully", product);
+  res.status(200).json({ msg: product });
 };
 
 const DeleteProduct = async (req, res, next) => {
   const id = req.params.id;
-
+  const productExist = await ProductModel.findOne({ _id: id });
+  if (!productExist) {
+    return res.status(404).json({ msg: "No Product Exist To Delete" });
+  }
   const product = await ProductModel.deleteOne({ _id: id });
-  return res.success("Delted Successfully")
+  res.status(200).json({ msg: product });
+  next();
 };
+
+const updateProduct= async(req,res,next)=>{
+  try {
+    const id=req.params.id;
+    const product=await ProductModel.findByIdAndUpdate(id,req.body);
+    if(!product){
+      return res.status(404).json({msg:"No Id Found To Update "})
+    }
+    return res.status(200).json({msg:product})
+    
+  } catch (error) {
+   return res.status(500).json({msg:error.message}) 
+  }
+}
+
+
 
 module.exports = {
   getById,
   CreateProduct,
-  getAllProductsTesting,
+  getAllProducts,
   DeleteProduct,
+  updateProduct
 };
